refactor(TipInput): extract TipInputPopup
